refactor(gateway): make index.ts a thin argv shim with dynamic role imports
